ttools: treat logs of negative values correctly in plots
Because I currently treat graphics coordinates as integer values in plot2,
graphics positions corresponding to logs of negative values were getting
calculated as NaNs and then turned into integer zero values by casting
NaN to int.  Although most such NaN coordinates were thrown out before
plotting, some, such as error bars terminating below zero on logarithmic
axes, were getting shown (since in this case Surface.dataToGraphics is
called with visibleOnly false), in obviously anomalous places.

Fix the code to check for NaNs before int casting is done.

This leaves the somewhat related problem that negative sense error bars
that would extend to negative values will not be displayed if the
relevant axis is logarithmic.  Probably it would be better to extend
them down infinitely.  This doesn't look very easy to fix; wait until
somebody complains till I worry about it.

+++ ttools/src/main/uk/ac/starlink/ttools/plot2/BasicTicker.java
@@ -400,7 +400,7 @@ public abstract class BasicTicker implements Ticker {
             Tick tick = ticks[ i ];
             String label = tick.getLabel();
             if ( label != null ) {
-                int gx = axis.dataToGraphics( tick.getValue() );
+                int gx = (int) axis.dataToGraphics( tick.getValue() );
                 Rectangle cbounds = captioner.getCaptionBounds( label );
                 AffineTransform oTrans =
                     orient.captionTransform( cbounds, cpad );
